InteractiveObjects

Documentation

getNewDoorsColors :: (Int, Int) -> Map -> [Color] -> [Color]

Function to get updated list of opened doors

isInsideButton :: Vector -> Map -> (Bool, Maybe (Int, Int))

Function to check if current player position is inside button

isInsideBattery :: Vector -> Map -> (Bool, Maybe (Int, Int))

Function to check if current player position is inside button

isInsideExit :: Vector -> Map -> (Bool, Maybe (Int, Int))

Function to check if current player position is inside exit

isInside :: Vector -> Map -> (GameObject -> Bool) -> (Bool, Maybe (Int, Int))