Maps

Documentation

data GameObject

Data type describing all possible objects in the game

Constructors

Wall

Door Color

Button Color

Floor

Entrance

Exit

Battery

Border

type Map = Array (Int, Int) GameObject

data Level

Level data type

Constructors

Level

Fields

levelMap :: [String]

openColors :: [Color]

initialPos :: Vector

initialDir :: Vector

parseMap :: [String] -> Map

Function to convert string representation of map into an array

symbolToObject :: Char -> GameObject

colorObject :: Char -> GameObject