MazeGenerator

Documentation

type Maze = Space String

prnt :: Show a => a -> a

Utility function for debugging

transposeMaze :: Maze -> Maze

Transpose maze

randomWallCoords :: RandomGen g => Area -> g -> (Coords, g)

Generating random coords for vertical wall. X coordinate shows the position of wall, coordinate y - the position of hole in the wall In order to prevent changes gluing and keep maze connected, walls are created on odd positions, ways on even.

createVerticalWall :: Coords -> Area -> String -> Maze

Create a Maze object for the wall, given coords, Area and object on the pass

allDoors :: [String]

head' :: [a] -> Maybe a

generateButtonRandomly :: RandomGen g => g -> Area -> String -> (Maze, g)

Function to generate button in random position.

deleteObjFromList :: [String] -> String -> [String]

clearExit :: Maze -> Maze

Function that guarantees that exit is available.

filterEmpty :: Maze -> Maze

Function to delete unneeded objects from Maze

genMaze :: RandomGen g => Area -> g -> [String] -> [String] -> Maze

Function to recursively generate maze. Every time maze is divided into 2 and each part is generated recursively

toUpper :: String -> String

Custom function to prevent Space naming conflicts

exit :: Coords -> Maze

Generation of exit coords.

generateRandomCoords :: RandomGen g => g -> Area -> (Coords, g)

Function to generate random coords in range

generateBatteriesList :: RandomGen g => g -> Maze -> [(Coords, String)] -> Int -> [(Coords, String)]

Function for recursive generation of batteries.

addBatteries :: RandomGen g => g -> Maze -> Int -> Maze

Function to generate batteries for appending

generateDifficultLevel :: RandomGen g => g -> Int -> Maze

Generate level with given difficulty

mazeToGrid :: [(Int, Int)] -> Maze -> [[String]] -> [[String]]

Function to convert maze to array

getObjectFromCoords :: (Int, Int) -> [(Coords, String)] -> String

Getter of object given coords.

getAreaFromMaze :: Maze -> Area

Getter of area given maze.

getMazeSize :: Maze -> Coords

Get maze size

getObjsFromMaze :: Maze -> [(Coords, String)]

Getter of objects from maze.

isObject :: Coords -> Maze -> Bool

Check if there is an object on the coords.

splitEvery :: Int -> [a] -> [[a]]

Function to split array into subarrays every n els.

joinString :: [[String]] -> String

Function for transforming Array of Strings to a single string

addBorders :: [String] -> [String]

Funtion to add borders to string representation of Maze